Moved statistic business rules to domain module
Fix #1932
