+++ pythonpro/cohorts/facade.py
@@ -3,6 +3,16 @@ from django.urls import reverse
 
 from pythonpro.cohorts.models import Cohort as _Cohort, CohortStudent, LiveClass as _LiveClass, Webinar as _Webinar
 
+__all__ = [
+    'get_all_cohorts_desc',
+    'find_cohort',
+    'find_most_recent_cohort',
+    'calculate_most_recent_cohort_path',
+    'find_webinars',
+    'find_webinar',
+    'find_live_class',
+]
+
 
 def get_all_cohorts_desc():
     return tuple(_Cohort.objects.order_by('-start'))
